dom/nodes/Node-cloneNode.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Node.cloneNode</title>
<link rel=help href="http://dvcs.w3.org/hg/domcore/raw-file/tip/Overview.html#dom-node-clonenode">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
function assert_equal_node(nodeA, nodeB) {
  assert_true(nodeB.attributes != nodeA.attributes);
  assert_true(nodeB.attributes !== nodeA.attributes);
  assert_equals(nodeB.attributes.length, nodeA.attributes.length);
  for (var i = 0, il = nodeA.attributes.length; i < il; ++i) {
    assert_not_equals(nodeB.attributes[i], nodeA.attributes[i]);
    assert_equals(nodeB.attributes[i].name, nodeA.attributes[i].name);
    assert_equals(nodeB.attributes[i].prefix, nodeA.attributes[i].prefix);
    assert_equals(nodeB.attributes[i].namespaceURI, nodeA.attributes[i].namespaceURI);
    assert_equals(nodeB.attributes[i].value, nodeA.attributes[i].value);
  }
}

test(function() {
  var el = document.createElement("foo");
  el.appendChild(document.createElement("bar"));
  el.setAttribute("a", "b");
  el.setAttribute("c", "d");
  var c = el.cloneNode(false);
  assert_equal_node(el, c);
  assert_equals(c.childNodes.length, 0);
})
</script>
